requirements.txt
# RNA 3D Structure Prediction - Requirements
# Python 3.8+

# Core scientific computing
numpy>=1.20.0
scipy>=1.7.0
pandas>=1.3.0

# Bioinformatics
biopython>=1.79

# Visualization (optional, for plotting results)
matplotlib>=3.4.0

# Testing (optional)
pytest>=6.2.0
pytest-cov>=2.12.0

# Development (optional)
black>=21.0
flake8>=3.9.0
mypy>=0.910

# Note: Install with:
#   pip install -r requirements.txt
#
# For development:
#   pip install -r requirements.txt black flake8 mypy
